+++ Bootstrap/Managers/Game.h
@@ -11,6 +11,7 @@ public:
 	static char* Name;
 	static char* UnityVersion;
 	static bool IsIl2Cpp;
+	static bool FirstRun;
 
 	static bool Initialize();
 	static bool SetupPaths();