+++ CHANGELOG.md
@@ -218,6 +218,7 @@
 194. Added Melon Hashes to Logs.   (Credits to TrevTV :P)
 195. Temporarily Disabled Analytics Blocker for x86 (32bit) to Fix Random Crash Issue.
 196. Fixed Issue with Console Coloring on Operating Systems older than Windows 10.
+197. Fixed Issue with app.info Does Not Exist Warning showing Twice.
 
 ---